pyproject.toml
# OmniBrain — Open-Source Personal AI Platform
# Built on Omnigent (universal agent framework)
# Modern Python packaging (PEP 621)

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omnibrain"
version = "0.1.0"
description = "Open-source AI platform that knows who you are, remembers everything, works 24/7, and grows through community-built Skills."
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Francesco Stabile" },
    { name = "Claude Opus 4.6" },
]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Office/Business",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Framework :: FastAPI",
    "Typing :: Typed",
]
keywords = [
    "ai-agent",
    "personal-ai",
    "ai-platform",
    "skill-protocol",
    "proactive",
    "local-first",
    "llm",
    "knowledge-graph",
    "open-source",
    "marketplace",
]

dependencies = [
    # ── Core (always required) ──
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.43",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",

    # ── API Server (platform core) ──
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pyjwt>=2.8.0",

    # ── Scheduling ──
    "apscheduler>=3.10.4",

    # ── Google APIs (for built-in Skills) ──
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
]

[project.urls]
Homepage = "https://github.com/FrancescoStabile/omnibrain"
Documentation = "https://github.com/FrancescoStabile/omnibrain#readme"
Repository = "https://github.com/FrancescoStabile/omnibrain"
Issues = "https://github.com/FrancescoStabile/omnibrain/issues"
Changelog = "https://github.com/FrancescoStabile/omnibrain/blob/main/CHANGELOG.md"

[project.optional-dependencies]
mcp = [
    "mcp[cli]>=1.26.0",
]
telegram = [
    "python-telegram-bot>=21.0",
]
chromadb = [
    "chromadb>=0.4.22",
]
secure = [
    "keyring>=25.0.0",
    "cryptography>=42.0.0",
]
local = [
    # For local embeddings (no OpenAI needed)
    "sentence-transformers>=3.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.9.0",
    "ruff>=0.3.0",
]
all = [
    "omnibrain[mcp]",
    "omnibrain[telegram]",
    "omnibrain[chromadb]",
    "omnibrain[secure]",
    "omnibrain[local]",
]

[project.scripts]
omnibrain = "omnibrain.__main__:main"
omnigent = "omnigent.__main__:main"
omnibrain-skill = "omnibrain.skill_cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/omnigent", "src/omnibrain"]
include = [
    "src/omnigent/**/*.md",
    "src/omnigent/**/*.txt",
    "src/omnibrain/**/*.md",
    "src/omnibrain/**/*.txt",
]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

# ══════════════════════════════════════════════════════════════════════════════
# Tool Configuration
# ══════════════════════════════════════════════════════════════════════════════

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
